Adding Shyp
Re: Remote opportunities - We are super interested in hiring great people who can't move to SF. We believe our processes can support remote work, but we understand this is a hard thing for us to assert with confidence. That said, we are extremely interested in building remote-work processes that scale - let's talk!